Further refactor code
